:sparkles: api: Check for publication before unfinalizing

+++ api/src/certification/session/domain/usecases/unfinalize-session.js
@@ -3,6 +3,7 @@
  * @typedef {import('../../../shared/domain/usecases/index.js').FinalizedSessionRepository} FinalizedSessionRepository
  */
 
+import { SessionAlreadyPublishedError } from '../errors.js';
 import { DomainTransaction } from '../../../../shared/domain/DomainTransaction.js';
 
 /**
@@ -11,6 +12,10 @@ import { DomainTransaction } from '../../../../shared/domain/DomainTransaction.j
  * @param {FinalizedSessionRepository} params.finalizedSessionRepository
  */
 const unfinalizeSession = async function ({ sessionId, sessionRepository, finalizedSessionRepository }) {
+  if (await sessionRepository.isPublished(sessionId)) {
+    throw new SessionAlreadyPublishedError();
+  }
+
   return DomainTransaction.execute(async (domainTransaction) => {
     await finalizedSessionRepository.delete({ sessionId, domainTransaction });
     await sessionRepository.unfinalize({ sessionId, domainTransaction });
